moving all files and setting go modules
