+++ comp/api/api/def/go.mod
@@ -1,0 +1,12 @@
+module github.com/DataDog/datadog-agent/comp/api/api/def
+
+go 1.21.0
+
+require go.uber.org/fx v1.21.1
+
+require (
+	go.uber.org/dig v1.17.1 // indirect
+	go.uber.org/multierr v1.10.0 // indirect
+	go.uber.org/zap v1.26.0 // indirect
+	golang.org/x/sys v0.0.0-20220412211240-33da011f77ad // indirect
+)
