+++ comp/api/api/impl/api_mock.go
@@ -5,7 +5,7 @@
 
 //go:build test
 
-package apiimpl
+package impl
 
 import (
 	"net"
