+++ pkg/util/grpc/context/go.mod
@@ -1,0 +1,3 @@
+module github.com/DataDog/datadog-agent/pkg/util/grpc/context
+
+go 1.21.0
